Time comparisons and visualiztaions
Time Comparisons of the iterative and recursive methids of pure birth trees are methodized in Time_comparison.py and a test run is shown in time_comp_results.py The iterative method is faster, especially for larger length limits and recursion depth errors also allpear after lingth limit of 900 in most cases.
An online visualization software https://itol.embl.de/ is used to visualize a splitting tree generated with speciation rate of 1.0 and length limit of 100. The pdf file is Tree_limit100_speciation1.pdf